Added leave-one-out validation for DFT comparison. Refactored directory structure.
